Rename 'frontend' folder to 'base'
