+++ assets/js/base/components/load-more-button/index.js
