+++ assets/js/base/components/read-more/utils.js
