+++ assets/js/base/components/reviews/index.js
