+++ assets/js/base/hocs/with-component-id.js
