+++ assets/js/blocks/reviews-by-product/frontend-block.js
@@ -8,10 +8,10 @@ import PropTypes from 'prop-types';
  * Internal dependencies
  */
 import { getOrderArgs, getReviews } from './utils';
-import LoadMoreButton from '../../frontend/components/load-more-button';
-import OrderBySelect from '../../frontend/components/order-by-select';
-import Reviews from '../../frontend/components/reviews';
-import withComponentId from '../../frontend/hocs/with-component-id';
+import LoadMoreButton from '../../base/components/load-more-button';
+import OrderBySelect from '../../base/components/order-by-select';
+import Reviews from '../../base/components/reviews';
+import withComponentId from '../../base/hocs/with-component-id';
 
 const enableReviewRating = !! ( typeof wc_product_block_data !== 'undefined' && wc_product_block_data.enableReviewRating );
